+++ packages/base-ui/rollup.config.mjs
@@ -13,7 +13,8 @@ const entries = [
   'src/icons/index.ts',
   'src/components/index.ts',
   'src/components/script/index.ts',
-  'src/components/server.ts'
+  'src/components/server.ts',
+  'src/components/shared.ts'
 ];
 
 const externalPackages = [